fix tests

+++ database/migrations/2014_10_12_000010_add_users_fields_table.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
 use \Xoco70\LaravelTournaments\DBHelpers;
 class AddUsersFieldsTable extends Migration
 {
@@ -12,8 +13,8 @@ class AddUsersFieldsTable extends Migration
     public function up()
     {
         Schema::table('users', function($table) {
-            $table->string('firstname');
-            $table->string('lastname');
+            $table->string('firstname')->default('firstname');
+            $table->string('lastname')->default('lastname');
 
         });
     }
@@ -26,10 +27,18 @@ class AddUsersFieldsTable extends Migration
     public function down()
     {
         DBHelpers::setFKCheckOff();
-        Schema::table('users', function($table) {
-            $table->dropColumn('firstname');
-            $table->dropColumn('lastname');
-        });
+        if (Schema::hasColumn('users', 'firstname')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('firstname');
+            });
+
+        }
+        if (Schema::hasColumn('users', 'lastname')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('lastname');
+            });
+
+        }
         DBHelpers::setFKCheckOn();
     }
 }
